third_party/WebKit/LayoutTests/animations/responsive/animations-responsive-textIndent.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id='container'>
    <div id='element'></div>
</div>
<script>

var container = document.getElementById('container');
var element = document.getElementById('element');

test(function() {
    container.style.fontSize = '10px';

    var keyframes = [
        {textIndent: '10em hanging'},
        {textIndent: '10em hanging'}
    ];
    var player = element.animate(keyframes, 10);
    player.pause();
    player.currentTime = 5;

    var textIndent = getComputedStyle(element).textIndent;
    container.style.fontSize = '20px';
    assert_not_equals(getComputedStyle(element).textIndent, textIndent);
}, 'textIndent responsive to style changes');

</script>
